Move window code to its own file
