+++ crates/wolf_engine_window/src/lib.rs
@@ -53,16 +53,12 @@
 //! This crate doesn't provide its own rendering functions.  Instead, it implements
 //! [`raw_window_handle`] traits in order for compatibility with external rendering libraries.
 
-use std::sync::Arc;
-
 use event::EventLoopBuilder;
-use winit::{
-    dpi::PhysicalSize,
-    window::{Window as WinitWindow, WindowAttributes},
-};
 
 mod context;
 pub use context::*;
+mod window;
+pub use window::*;
 
 pub mod event;
 pub use uuid::Uuid;
@@ -75,122 +71,6 @@ pub mod raw_window_handle;
 /// Initialize the window system.
 pub fn init() -> EventLoopBuilder {
     EventLoopBuilder::new()
-}
-
-/// The settings used by the [`WindowContext`] when creating the window.
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct WindowSettings {
-    pub title: String,
-    pub size: (u32, u32),
-    pub is_resizable: bool,
-    pub is_visible: bool,
-}
-
-impl WindowSettings {
-    /// Set the window's title.
-    pub fn with_title(mut self, title: &str) -> Self {
-        self.title = title.to_string();
-        self
-    }
-
-    /// Set the window's size.
-    pub fn with_size(mut self, size: (u32, u32)) -> Self {
-        self.size = size;
-        self
-    }
-
-    /// Set whether or not the window should be resizable.
-    pub fn with_resizable(mut self, is_resizable: bool) -> Self {
-        self.is_resizable = is_resizable;
-        self
-    }
-
-    /// Set whether or not the window should be visible.
-    pub fn with_visible(mut self, is_visible: bool) -> Self {
-        self.is_visible = is_visible;
-        self
-    }
-}
-
-impl Default for WindowSettings {
-    fn default() -> Self {
-        Self {
-            title: "Untitled".to_string(),
-            size: (1280, 720),
-            is_resizable: true,
-            is_visible: true,
-        }
-    }
-}
-
-impl From<WindowSettings> for WindowAttributes {
-    fn from(val: WindowSettings) -> Self {
-        WindowAttributes::default()
-            .with_title(val.title)
-            .with_inner_size(PhysicalSize::new(val.size.0, val.size.1))
-            .with_resizable(val.is_resizable)
-            .with_visible(val.is_visible)
-    }
-}
-
-/// A window.
-#[derive(Clone, Debug)]
-pub struct Window {
-    uuid: Uuid,
-    inner: Arc<WinitWindow>,
-}
-
-impl PartialEq for Window {
-    fn eq(&self, other: &Self) -> bool {
-        self.uuid == other.uuid
-    }
-}
-
-impl Eq for Window {}
-
-impl Window {
-    pub(crate) fn new(inner: Arc<WinitWindow>) -> Self {
-        Self {
-            uuid: Uuid::new_v4(),
-            inner,
-        }
-    }
-
-    pub fn id(&self) -> Uuid {
-        self.uuid
-    }
-
-    /// Get the current size of the window.
-    pub fn size(&self) -> (u32, u32) {
-        let size = self.inner.inner_size();
-        (size.width, size.height)
-    }
-}
-
-impl rwh_06::HasWindowHandle for Window {
-    fn window_handle(&self) -> Result<rwh_06::WindowHandle<'_>, rwh_06::HandleError> {
-        rwh_06::HasWindowHandle::window_handle(&self.inner)
-    }
-}
-
-impl rwh_06::HasDisplayHandle for Window {
-    fn display_handle(&self) -> Result<rwh_06::DisplayHandle<'_>, rwh_06::HandleError> {
-        rwh_06::HasDisplayHandle::display_handle(&self.inner)
-    }
-}
-
-#[cfg(feature = "rwh_05")]
-unsafe impl rwh_05::HasRawWindowHandle for Window {
-    fn raw_window_handle(&self) -> rwh_05::RawWindowHandle {
-        rwh_05::HasRawWindowHandle::raw_window_handle(&self.inner)
-    }
-}
-
-#[cfg(feature = "rwh_05")]
-unsafe impl rwh_05::HasRawDisplayHandle for Window {
-    fn raw_display_handle(&self) -> rwh_05::RawDisplayHandle {
-        rwh_05::HasRawDisplayHandle::raw_display_handle(&self.inner)
-    }
 }
 
 #[cfg(test)]
